This is an attempt to address #1662. Working on a code base that makes extensive us of zip archives containing parquet files that need to be read on the fly to load data, I noticed that `find` was a major bottleneck in that process.

I have made a zip specific implementation of `find` here that relies on the file list of the zip file, rather than explicitly walking.

Here is how I measured the performance:

```
from fsspec.implementations.zip import ZipFileSystem

# Example achieve with roughly 900 files, in some deep directory structures
file_system = ZipFileSystem("example.zip")
%timeit file_system.find("/")
```

Performance for find on current `master`
```
2.14 s ± 146 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Performance for find with this fix applied:
```
318 µs ± 11.5 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

Let me know what you think.