@johandahlberg 2024-08-21
Thank you for the review @martindurant! I refactored the code quite a bit, to make use of the `dir_cache` instead of directly accessing the zip file. I think it made the code easier to follow.

Regarding removing the `maxdepth` that would be a change in the behavior compared to the `find` method on the `AbstractFileSystem` - since my initial strategy here was to try to implement something that had parity in the output I kept it in. If you still want me to remove it let me know and I'll do it (or feel free to push any changes you want to this branch if you prefer that).

I implemented it as suggested by you now, but just filtering for it at the end if it is set.